# apps/main-service/src/domain/user/dto/update-user.dto.ts
import { IsEmail, IsOptional, IsString } from 'class-validator';
import { ApiProperty } from '@nestjs/swagger';

export class UpdateUserDto {
  @ApiProperty({ example: 'jake@jake.jake', required: false })
  @IsOptional()
  @IsEmail()
  email?: string;

  @ApiProperty({ example: 'I like to skateboard', required: false })
  @IsOptional()
  @IsString()
  bio?: string;

  @ApiProperty({ example: 'https://i.stack.imgur.com/xHWG8.jpg', required: false })
  @IsOptional()
  @IsString()
  image?: string;
}